% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copyInputs.R
\name{copyInputs}
\alias{copyInputs}
\title{Copy input files from WorldPop server to your local working directory}
\usage{
copyInputs(srcdir, outdir, overwrite = F, OS.type = "windows")
}
\arguments{
\item{srcdir}{Directory of source files}

\item{outdir}{Directory to write output files to}

\item{overwrite}{Logical indicating to overwrite local files if they already exist}

\item{OS.type}{Type of operating system retreived from \code{.Platform$OS.type}}
}
\value{
Writes input files to disk
}
\description{
Copy input files from WorldPop server to your local working directory
}
