% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDRC.R
\name{dataDRC}
\alias{dataDRC}
\title{Creates jags data for the DRC analysis}
\usage{
dataDRC(indir = "in/", outdir = "out/drc/", seed = 42)
}
\arguments{
\item{indir}{Directory with input data}

\item{outdir}{Directory where to store output data}

\item{seed}{Random seed}
}
\value{
JAGS data objects are saved to disk
}
\description{
Creates jags data for the DRC analysis
}
