% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSim.R
\name{dataSim}
\alias{dataSim}
\title{Simulate population, collect samples, and create jags data}
\usage{
dataSim(sampling = "weighted", model_weights = T, n.weighted = 500,
  n.random = 500, n.real = 10000, maxarea = 10, beta = 0.2,
  type1_prop = 0.5, med1 = 300, sigma1 = 150, med2 = 100,
  sigma2 = 50, outdir = "sim", seed = 42)
}
\arguments{
\item{sampling}{Type of sampling ("weighted" or "random"; Default = "weighted").}

\item{model_weights}{Toggle model weights (T or F; Default = T).}

\item{n.weighted}{Number of population weighted samples (Default = 500).}

\item{n.random}{Number of random samples (Default = 500).}

\item{n.real}{Total number of spatial units in the population (Default = 1e4).}

\item{maxarea}{The maximum settled area possible for a spatial unit (Default = 10).}

\item{beta}{Effect of simulated covariate on population density (Default = 0.2).}

\item{type1_prop}{Proportion of spatial units that are settlement type 1 (Default = 0.5).}

\item{med1}{Median population density for settlement type 1 (Default = 300).}

\item{sigma1}{Standard deviation of population densities within settlement type 1 (Default = 150).}

\item{med2}{Median population density for settlement type 2 (Default = 100).}

\item{sigma2}{Standard deviation of population densities within settlement type 2 (Default = 50).}

\item{outdir}{Name of the output directory to create for simulation results (Default = "sim").}

\item{seed}{Random seed (Default = 42).}
}
\value{
All results are saved to disk in the designated output folder.
}
\description{
Simulate population, collect samples, and create jags data
}
