% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataZMB.R
\name{dataZMB}
\alias{dataZMB}
\title{Creates jags data for the DRC analysis}
\usage{
dataZMB(indir = "in/", outdir = "out/zmb/", area = F, seed = 42)
}
\arguments{
\item{indir}{Directory with input data}

\item{outdir}{Directory where to store output data}

\item{area}{Logical indicating to include EA settled areas in JAGS data. If FALSE, areas = 1.}

\item{seed}{Random seed}
}
\value{
JAGS data objects are saved to disk
}
\description{
Creates jags data for the DRC analysis
}
