% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inits.R
\name{inits}
\alias{inits}
\title{Create initials for MCMC chains in JAGS}
\usage{
inits(dat, chains = 3)
}
\arguments{
\item{dat}{JAGS data object}

\item{chains}{Number of MCMC chains that will be run}
}
\value{
A list of lists. Each sub-list includes an item for each parameter with initials
}
\description{
Create initials for MCMC chains in JAGS
}
