% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jagsModel.R
\name{jagsModel}
\alias{jagsModel}
\title{Fit a jags model using simulation results}
\usage{
jagsModel(dir, toggleCov = T)
}
\arguments{
\item{dir}{Directory where simulation results are located (required)}

\item{toggleCov}{Toggle inclusion of covariate in the model (Default = T)}
}
\value{
Outputs are saved to disk in the designated directory. Outputs include the jags model (./objects/jm.rds), trace plots (./trace/*), and a data frame with posterior samples for monitored parameters (./objects/d.rds)
}
\description{
Fit a jags model using simulation results
}
