% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapKinshasa.R
\name{mapKinshasa}
\alias{mapKinshasa}
\title{Create a map of Kinshasa study area}
\usage{
mapKinshasa(file)
}
\arguments{
\item{file}{Output file to save map}
}
\value{
Map is saved to disk
}
\description{
Create a map of Kinshasa study area
}
