% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_to_df.R
\name{mcmc_to_df}
\alias{mcmc_to_df}
\title{Convert an mcmc list to a data frame}
\usage{
mcmc_to_df(x)
}
\arguments{
\item{x}{MCMC list object}
}
\value{
Data frame with a column for each parameter and a row for each mcmc sample
}
\description{
Convert an mcmc list to a data frame
}
