% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_to_median.R
\name{mean_to_median}
\alias{mean_to_median}
\title{Convert log-normal mean to median (equations derived from ?rlnorm)}
\usage{
mean_to_median(mu, log_sigma, inverse = F)
}
\arguments{
\item{mu}{Mean of log-normal distribution. If inverse = T, this argument is the median.}

\item{log_sigma}{Standard deviation of log-normal distribution}

\item{inverse}{Logical indicating if inverse should be calculated (i.e. convert a median to a mean)}
}
\value{
If inverse = F, it returns the median of the log-normal distribution. If inverse = T, it returns the mean of the log-normal.
}
\description{
Convert log-normal mean to median (equations derived from ?rlnorm)
}
