% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotModel.R
\name{plotModel}
\alias{plotModel}
\title{Create a plot of probability distributions for data and model. If the true distribution is known from simulations, that distribution can be plotted as well.}
\usage{
plotModel(jd, d, real1 = NA, real2 = NA, mar = c(4.5, 2.5, 1, 1),
  legend = F, xlab = NA, xaxt = "s", xmax = 1000, main = NA,
  line.main = -1.5, adj.main = 0.5)
}
\arguments{
\item{jd}{JAGS data object}

\item{d}{JAGS model mcmc list saved as a data frame (see ?mcmc_to_df)}

\item{real1}{Optional vector of true population sizes for each spatial unit in settlement type 1}

\item{real2}{Optional vector of true population sizes for each spatial unit in settlement type 2}

\item{mar}{Plot margins for use with par()}

\item{legend}{Logical indicating to include plot legend}

\item{xlab}{Label for x-axis}

\item{xaxt}{X axis style; include it or not c('s','n')}

\item{xmax}{Maximum value for x-axis}

\item{main}{Title for plot}

\item{line.main}{Line to position plot title}

\item{adj.main}{Adjustment factor for plot title (0=left, 0.5=middle, 1=right)}
}
\value{
Returns a plot object
}
\description{
Create a plot of probability distributions for data and model. If the true distribution is known from simulations, that distribution can be plotted as well.
}
