% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotModelPanel.R
\name{plotModelPanel}
\alias{plotModelPanel}
\title{Creates a multipanel plot using plotModel()}
\usage{
plotModelPanel(file, sims, dir, plotReal = F, xmax = 1000)
}
\arguments{
\item{file}{File name to save jpeg image of plot}

\item{sims}{The simulation names to be included in the plot (must match directory names for simulation results)}

\item{dir}{Root directory containing simulation results}

\item{plotReal}{Logical to include the true population sizes in the plot}

\item{xmax}{Maximum value for the x-axis in plots}
}
\value{
Writes the plot to disk
}
\description{
Creates a multipanel plot using plotModel()
}
