% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTotals.R
\name{plotTotals}
\alias{plotTotals}
\title{Creates a barplot with population totals}
\usage{
plotTotals(dat, file, pos.legend = "topright", plotReal = F,
  main = "Population Totals")
}
\arguments{
\item{dat}{Plot data (see ?plotTotalsData)}

\item{file}{File name of jpeg image to save plot}

\item{pos.legend}{Position of legend}

\item{plotReal}{Logical to include plot legend}

\item{main}{Plot title}
}
\value{
Saves plot to disk as jpeg
}
\description{
Creates a barplot with population totals
}
