% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTotalsData.R
\name{plotTotalsData}
\alias{plotTotalsData}
\title{Plot population totals estimated from each simulated sampling design}
\usage{
plotTotalsData(dir, alpha = 0.05, plotReal = F)
}
\arguments{
\item{dir}{Directory for where simulation results are located (required)}

\item{alpha}{Parameter to define confidence limit (Default = 0.05 for 95% confidence intervals)}

\item{plotReal}{Toggle a line to show the true population size from simulations (Default = F)}
}
\value{
Results are saved to disk as jpeg file.
}
\description{
Plot population totals estimated from each simulated sampling design
}
