% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predPop.R
\name{predPop}
\alias{predPop}
\title{Posterior predictions from Bayesian model using fitted parameter values}
\usage{
predPop(sim, d, t)
}
\arguments{
\item{sim}{Population simiulation object (see ?simPop)}

\item{d}{JAGS model saved as a data frame (see ?mcmc_to_df)}

\item{t}{Settlement type where predictions are being made (t = 1 or 2)}
}
\value{
Returns a matrix containig a vector of posterior predictions for each location (columns represent locations, rows represent posterior samples)
}
\description{
Posterior predictions from Bayesian model using fitted parameter values
}
