% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlnorm_trunc.R
\name{rlnorm_trunc}
\alias{rlnorm_trunc}
\title{Truncated log-normal distribution}
\usage{
rlnorm_trunc(n, m, s, trunc.val = Inf)
}
\arguments{
\item{n}{Number of samples}

\item{m}{Median of log-normal distribution (see ?rlnorm)}

\item{s}{Standard deviation of log-normal distribution (see ?rlnorm)}

\item{trunc.val}{Value for truncating the distribution (define the maximum possible value)}
}
\value{
Vector of random samples from the distribution
}
\description{
Truncated log-normal distribution
}
