% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_to_log.R
\name{sig_to_log}
\alias{sig_to_log}
\title{Convert sigma to log_sigma for log-normal distribution (equations derived from ?rlnorm)}
\usage{
sig_to_log(med, sigma)
}
\arguments{
\item{med}{Median of log-normal distribution}

\item{sigma}{Standard deviation of log-normal distribution (on natural scale)}
}
\value{
Standard deviation of the log-normal distribution (on the log scale)
}
\description{
Convert sigma to log_sigma for log-normal distribution (equations derived from ?rlnorm)
}
