% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simPop.R
\name{simPop}
\alias{simPop}
\title{Simulate a population}
\usage{
simPop(n = 1000, med = 100, sigma = 50, beta = 0.2, maxarea = 10,
  maxpop = Inf)
}
\arguments{
\item{n}{Number of spatial units}

\item{med}{Median of population densities}

\item{sigma}{Standard deviation of population densities}

\item{beta}{Effect of covariate on population density}

\item{maxarea}{Maximum settled area possible for a spatial unit}

\item{maxpop}{Absolute maximum population density used to truncate a log-normal}
}
\value{
A list object with an item for population count (N), density (D), settled area (A), covariate (x), and simulation parameters (med, sigma, beta)
}
\description{
Creates a list object with population sizes, population densities, settled areas, and covariate values for every spatial unit
}
