% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights_calc.R
\name{weights_calc}
\alias{weights_calc}
\title{Calculate model weights from population numbers for weighted-likelihood model}
\usage{
weights_calc(numerator1, denominator, numerator2 = NULL, inverse = F)
}
\arguments{
\item{numerator1}{Vector of population sizes to use as weights}

\item{denominator}{Denominator to scale weights}

\item{numerator2}{Vector of population sizes for unweighted (random) samples to combine with weighted samples (numerator1)}

\item{inverse}{Logical indicating if inverse weights should be returned}
}
\value{
Vector of weights
}
\description{
Calculate model weights from population numbers for weighted-likelihood model
}
